BUG FIXED. Found vulnerabilities in npm package dependencies
Errors:
- Gets on GitHub 4 dependabot alerts: 'set-value'
- Running the command 'npm audit' in the terminal returns a list with 'Some vulnerabilities require your attention to resolve':
glob-parent --- High --- Regular expression denial of service
set-value ---- High ---- Prototype Pollution in set-value

References:
https://stackoverflow.com/a/69591894
- npm-force-resolutions not working when installing a new package
https://stackoverflow.com/a/67446313

Solution:
- Added each vulnerability of the dependency of dependency to the 'resolutions’ in the ‘package.json’ file.
Each vulnerability is replaced with a different package or with a more stable version of the package.
- Created npm 'preinstall' script to fix the vulnerabilities even if the 'package-lock.json' file has been deleted or does not exist yet.
This command first does the normal npm installation to create the 'package-lock.json' file and then updates the dependency resolution to the mentioned version.